requirements.txt
accelerate==1.7.0
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.7
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.52.2
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
audioop-lts==0.2.1
av==14.4.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
build==1.2.2.post1
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.0.12
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.4
ctranslate2==4.6.0
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
deepgram-sdk==2.12.0
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==1.16.0
duckduckgo_search==8.1.1
durationpy==0.10
ecdsa==0.19.1
effdet==0.4.1
emoji==2.14.1
faiss-cpu==1.12.0
fastapi==0.115.9
faster-whisper==1.1.1
ffmpeg==1.4
filelock==3.18.0
filetype==1.2.0
firebase==4.0.1
firebase_admin==7.1.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.6.2
fsspec==2025.5.1
fuzzywuzzy==0.18.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-cloud-vision==3.10.2
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
groq==0.29.0
grpcio==1.72.1
grpcio-status==1.71.0
grpcio-tools==1.72.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.27
langchain-anthropic==0.3.15
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.0.10
langchain-groq==0.3.5
langchain-huggingface==0.3.0
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langcodes==3.5.0
langdetect==1.0.9
langgraph==0.6.2
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.2
langgraph-sdk==0.2.0
langsmith==0.3.45
language_data==1.3.0
llvmlite==0.44.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mlxtend==0.23.4
mmh3==5.1.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.4.4
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.41.0
neo4j==5.28.2
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.2.6
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.0
openai==1.100.2
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
opencv-python==4.11.0.86
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pdf2image==1.17.0
pi_heif==1.0.0
pikepdf==9.9.0
pillow==11.2.1
postgrest==1.1.1
posthog==4.2.0
preshed==3.0.10
primp==0.15.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pymongo==3.12.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==6.0.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
realtime==2.7.0
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
sqlite-vec==0.1.6
srsly==2.5.1
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.45.3
storage3==0.12.1
streamlit==1.45.1
streamlit-webrtc==0.63.3
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.16
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unicorn==2.1.3
unstructured-client==0.37.2
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
uritemplate==4.2.0
urllib3==2.4.0
uv==0.8.13
uvicorn==0.34.3
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.5
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
langchain_mcp_adapters